SLING-1761 : JcrPropertyResource sets incorrect content length for strings containing non-ascii character Applied slightly modified patch from Julian Sedding
git-svn-id: https://svn.apache.org/repos/asf/sling/trunk@1053879 13f79535-47bb-0310-9956-ffa450edef68

+++ bundles/jcr/resource/src/test/java/org/apache/sling/jcr/resource/internal/helper/jcr/JcrItemResourceTestBase.java
@@ -82,7 +82,7 @@ public class JcrItemResourceTestBase extends RepositoryTestBase {
         } else {
             try {
                 for (byte b : expected) {
-                    assertEquals(b, actual.read());
+                    assertEquals(b, (byte)actual.read());
                 }
             } finally {
                 try {
